Move SlotId to LPC55-specific API
