+++ drv/lpc55-update-api/src/lib.rs
@@ -79,6 +79,34 @@ pub enum UpdateTarget {
     Bootloader = 4,
 }
 
+/// Designates a firmware image slot in parts that have fixed slots (rather than
+/// bank remapping).
+///
+/// This `enum` is used as part of the wire format for SP-RoT communication, and
+/// therefore cannot be changed at will; see discussion in `drv_sprot_api::Msg`
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Eq,
+    PartialEq,
+    FromPrimitive,
+    Serialize,
+    Deserialize,
+    SerializedSize,
+)]
+pub enum SlotId {
+    A,
+    B,
+}
+
+impl TryFrom<u16> for SlotId {
+    type Error = ();
+    fn try_from(i: u16) -> Result<Self, Self::Error> {
+        Self::from_u16(i).ok_or(())
+    }
+}
+
 // This value is currently set to `lpc55_romapi::FLASH_PAGE_SIZE`
 //
 // We hardcode it for simplicity, and because we cannot,and should not,
